requirements.txt
accelerate==0.15.0
diffusers==0.9.0
fastapi==0.88.0
Pillow==9.3.0
streamlit==1.15.2
torch==1.13.0
transformers==4.25.1
uvicorn[standard]==0.20.0